+++ src/main/java/com/example/CorporateLearningManagmentSystem/repository/Enrolled.java
@@ -2,5 +2,5 @@ package com.example.CorporateLearningManagmentSystem.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface Enrolled extends JpaRepository<com.example.CorporateLearningManagmentSystem.entity.Enrolled,Long> {
+public interface Enrolled extends JpaRepository<com.example.CorporateLearningManagmentSystem.entity.Enrolled,Integer> {
 }
