+++ src/main/java/com/example/CorporateLearningManagmentSystem/repository/ModuleRepository.java
@@ -3,5 +3,5 @@ package com.example.CorporateLearningManagmentSystem.repository;
 import com.example.CorporateLearningManagmentSystem.entity.Module;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ModuleRepository extends JpaRepository<Module, Long> {
+public interface ModuleRepository extends JpaRepository<Module, Integer> {
 }
