+++ src/main/java/com/example/CorporateLearningManagmentSystem/repository/ResourceRepository.java
@@ -3,5 +3,5 @@ package com.example.CorporateLearningManagmentSystem.repository;
 import com.example.CorporateLearningManagmentSystem.entity.Resource;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ResourceRepository extends JpaRepository<Resource, Long> {
+public interface ResourceRepository extends JpaRepository<Resource, Integer> {
 }
